pyproject.toml
[project]
name = "mcp-benchmark-stack"
version = "0.1.0"
description = "Real MCP benchmark stack for security evaluation"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "anyio>=4.0.0",
    "starlette>=0.41.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["client"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
